refactor(server): provider-symmetric webhook sub-packages + 3 correctness fixes
Round 3 review surfaced four real issues. Each fix below is one of them.

1. Resilience4j Retry would burn 5 attempts on an `IllegalArgumentException`
   thrown by `JetStreamPublisher.streamFor()` (which only fires on a
   programmer bug — unknown subject prefix — but every "retry storm"
   counter increment skews the SLI). Add
   `ignoreExceptions(IllegalArgumentException.class)` to the Retry config.

2. Both controllers were INFO-logging every successful publish
   ("Published GitLab webhook: subject=… event=…"). The publisher already
   DEBUG-logs the same fact. At fleet scale ≈ 10 RPS that's ~864k lines
   per pod per day with no extra signal. Demote both to DEBUG.

3. Provider sub-packages were asymmetric — `webhook/github/` held the
   GitHub controller + subject builder, but the GitLab controller lived
   in `webhook/web/` alongside `WebhookPayloadSizeFilter`, and the GitLab
   subject builder + token verifier + the GitHub-only `HmacVerifier` all
   sat naked in the root `webhook/` package. A newcomer looking for the
   GitLab controller looked under `webhook.gitlab.*` first and was wrong.
   New layout:
     webhook/gitlab/  GitLabWebhookController, GitLabSubjectBuilder,
                     GitLabTokenVerifier
     webhook/github/  GitHubWebhookController, GitHubSubjectBuilder,
                     HmacVerifier  (HmacVerifier is GitHub-only, not shared)
     webhook/web/     WebhookPayloadSizeFilter (cross-cutting HTTP layer)
     webhook/        DedupIdResolver, JetStreamPublisher, StreamBootstrap,
                     PublishRequest, WebhookConfiguration,
                     WebhookGracefulShutdown, WebhookHealthIndicator
   Tests moved to match. Stale `target/classes` from the move caused a
   ConflictingBeanDefinition on first re-run; cleaned via `mvn clean`.

4. Polish: `WebhookProperties` had `@DurationUnit(...) @DefaultValue("9s")`
   on every Duration field. The suffix on the default wins, so the unit
   annotation is silent noise on every field. Drop the redundant
   `@DurationUnit` annotation across all five Duration sites and the now-
   unused imports. Dead `effectiveEvent != null` guard in
   `GitLabWebhookController` (the prior assignment guarantees non-null)
   reduced to `!effectiveEvent.isEmpty()`.

2986 unit/architecture + 11 integration tests pass.
