+++ server/src/main/java/de/tum/cit/aet/hephaestus/gitprovider/webhook/gitlab/GitLabSubjectBuilder.java
@@ -1,4 +1,4 @@
-package de.tum.cit.aet.hephaestus.gitprovider.webhook;
+package de.tum.cit.aet.hephaestus.gitprovider.webhook.gitlab;
 
 import java.util.ArrayList;
 import java.util.List;
