+++ server/src/test/java/de/tum/cit/aet/hephaestus/gitprovider/webhook/github/HmacVerifierTest.java
@@ -1,4 +1,4 @@
-package de.tum.cit.aet.hephaestus.gitprovider.webhook;
+package de.tum.cit.aet.hephaestus.gitprovider.webhook.github;
 
 import static org.assertj.core.api.Assertions.assertThat;
